requirements.txt
numpy==1.24.3
pandas==1.5.3
cx_Oracle==8.3.0
python-dotenv==1.0.0
openpyxl==3.1.2